Fixed percentage missing from css width on progress bar update, Added $set flag to mongodb update
Added the $set flag, so that the field is added to the document,
instead of removing existing fields within the document. This will
prevent the whole document from being updated.

The ‘%’ sign appended to the number value while setting the width
should update the progress bar as the file is uploaded